feat: refactor remote server entry points to unify transport handling and remove Iroh-specific functions
