+++ crates/burn-remote/src/server/spawn.rs
@@ -15,3 +15,30 @@ where
 {
     wasm_bindgen_futures::spawn_local(future);
 }
+
+/// Resolve when the process is asked to stop (Ctrl+C, or `SIGTERM` on Unix).
+///
+/// The single shutdown trigger shared by the turnkey WebSocket and Iroh server entry points.
+#[cfg(all(not(target_family = "wasm"), any(feature = "websocket", feature = "iroh")))]
+pub(crate) async fn os_shutdown_signal() {
+    let ctrl_c = async {
+        tokio::signal::ctrl_c()
+            .await
+            .expect("failed to install Ctrl+C handler");
+    };
+
+    #[cfg(unix)]
+    let terminate = async {
+        tokio::signal::unix::signal(tokio::signal::unix::SignalKind::terminate())
+            .expect("failed to install signal handler")
+            .recv()
+            .await;
+    };
+    #[cfg(not(unix))]
+    let terminate = std::future::pending::<()>();
+
+    tokio::select! {
+        _ = ctrl_c => {},
+        _ = terminate => {},
+    }
+}
